+++ packages/std/src/iterator.rs
@@ -6,9 +6,6 @@ use crate::errors::StdError;
 /// allows contracts to reuse the type when deserializing database records.
 pub type Record<V = Vec<u8>> = (Vec<u8>, V);
 
-#[deprecated(note = "Renamed to Record, please use this instead")]
-pub type Pair<V = Vec<u8>> = Record<V>;
-
 #[derive(Copy, Clone)]
 // We assign these to integers to provide a stable API for passing over FFI (to wasm and Go)
 pub enum Order {